Directly transition to first_sync instead of invoking action
Based on our previous spelunking (as well as reported errors in Sentry), we currently have a race condition where the first route
transition has not always completed when we attempt to call an action on the `ApplicationRoute`.

Given the `renderFirstSync` action simply redirects, we should be able to use the (currently private) `routing` service as an interim replacement.

If the exception is what is keeping users from proceeding through the typical auth flow, this should hopefully address it.