[TEMPORARY]: Avoid triggering error-prone event
This was the main thing keeping us from upgrading, as the (private?)
behavior of the router appears to have changed.

i.e. it used to be possible to call `send` on the return value of
Ember.getOwner(this).lookup('router:main') -- which returned an instance
of the Router.

This now errors out.